sim: fix compile warning in sigint.h
Resolve the following compiler warning by calling the PeNSSignals
constructor directly rather than relying on an initializer-list
conversion syntax.

src/sigint.h:163:40: warning: suggest braces around initialization of
      subobject [-Wmissing-braces]
        PeNSSignals pull_ns() {return {0};}
                                       ^
                                       {}
1 warning generated.

PeNSSignals is made up of entirely POD and so the default constructor
on it would be to initialize all members to zero.  So, the previous
syntax has the same behavior as relying on the default constructor.

Signed-off-by: Patrick Williams <alpawi@amazon.com>