TCC's compile_convolve api takes in the argument output data type (o_dtype).
Previously, output data type = upcast(input data type), i.e. whatever was
the accumulated result in the psum engine became the output data type.
We want to preserve this option for debugging the psum engine, but we
also want to have the option to convert the type for easier processing
of the next layer in the state buffer.