Allow any shape for Const layer
Earlier Const layer shape had form [1,*,1,1]. That is too strict for an arbitrary
Const layer - that applies for Const layer that is input to BiasAdd.
Now the assert is for the Const layer as the input for BiasAdd, not in general.

Regressions pass - Jindrich needs more general Const layers, and he will verify that
new code satisfies his needs.