Change EventId_Invalid from (-1) to (NUM_TPB_EVENTS-1).
For some reason Inkling uses event_id even when event_mode==0.
When I tried my experiments with event field of 16b (instead of 8b)
and with NUM_TPB_EVENTS==2048, invalid id (-1) was interpreted as 0xFFFF,
and that was beyound the range of 2048.
Q: Should Inkling do anything with event_id when event_mode==0?
Also, initialize event_id in the instructions to 0.