Increase num weights in LdWeights for alignment.
When the weights are loaded with delta=-1, the last weight must
be aligned with byte 7 in an octet, i.e., if LA == last weight's address,
then LA == 8*N+7 for int8, and LA == 8*N+6 for int16 or fp16.
That is part of the performance check in new ISA.
The additional junk weights do not affect the result since only real
OFMAP columns will be collected.

Revert wait+clear to use WAIT;CLEAR rather than either
WAIT(mode=wait+clear) or NOP(wait-mode=wait+clear), as the latter two
appear not to work in Inkling.