Fix 8-resnet50_fp32_wave regression failure.
tffe script has --images optional argument to take image file names
as list but the default value is None. As part of a code review
comment, I changed the default value of args.images to []
from None(commit 9e84c5f8). However, inside the script the
code compares args.images with None to take different code path,
which was not converted. Because of this 8-resnet50_fp32_wave
test fails and I didnt notice it because this test is waived.

Reverting the change to fix this problem.

cr https://code.amazon.com/reviews/CR-1917014