build: remove direct 'make' invocations
Calling 'make' directly prevents parallel builds
from working.  make typically overrides the MAKE
variable with the "location of the jobserver".  If
'make' is used instead of '$(MAKE)', the jobserver
cannot be found in sub-makes and make reverts to
a single-threaded build (and announces a warning).

Change all 'make' to '$(MAKE)'.

Change-Id: I64095c7deeb26db2826dce70bd1575f4f0e13030
Signed-off-by: Patrick Williams <alpawi@amazon.com>