Support for writing only final Ofmap save waveops
There are 3 types of Save waveops:
1. Final output - no out edges in wavegraph
2. Offload (to be loaded later) - out edges present
3. Debug (to be saved somewhere) - no out edges

RT does not support 2 and 3. ME has added a flag to distinguish 1 from other cases.
Currently we ignore 2 and 3.
This solves a problem where we were sending on of type 3 saves to $OUTPUT,
and the data coming over $OUTPUT descs had more bytes than expected.

Change-Id: I632aff13afd636771fc58ce650b7c13554fb28a5