Generate correct descriptors
Change-Id: Ia14711a51446365b6790218b65e893d37a3c00c4

Add NOP with 128 cycles in front of every DMA trigger

Pooling or other engine sets event (mode=2)
after the last element of the FIRST channel is written,
not after the last element of the LAST channel is written.

Therefore DMA can start before the last channel is written
so a very fast DMA could read the last channel before it is written completely.
Similarly, DMA could overwrite the last channel before it is consumed.
The latter is less likely to cause trouble because DMA Load starts
after an engine produces the last element of the first channel of its output
rather than after the engine consumes the last element of the first channel of its input.

Change-Id: Ifc9ecbd7dce15f571a9521cfc0077b5a06573180

Factor out DMA barrier.

+ DMA barrier was missing for few DmaTrigger Instructions

Change-Id: Ibff2cb2f38398ca56f6bf0d5dce8b4fbf2b69f9d