[jenkins] Use Ubuntu 16.04 in Jenkins.
Runtime uses grpc library which is precompiled on keana C5
instance(which runs Ubuntu 16.04). The precompilation is done
because grpc compile time is ~15min on C5 itself.

Using the precompiled libaries cause problem in Jenkins because
of old linker(2.24) in Jenkins. So upgrading Jenkins to use
the same image as C5.

Change-Id: I4e6d4e5618d07ea86115e244fe23f98eaf22dc8e