Add support for non compiler tests in qemu_rt
In this version of the change, qemu_rt will support running
non compiler test. Only running on Qemu has been tested with
this change and not zebu. Most of the flow is the same as
compiler test except the following
- krtd is used instead of krt_bin
- pytest is invoked from the qemu and that runs all the tests
- new class krtd is created which is child of krt. krt is
currently using kbin_rt and this will be removed once all
tests are maded to use krtd (compiler tests are using kbin_rt)
- new class pythontest is created which will run pytest
on the scripts that were copied as part of krt install

Change-Id: I9ebf2093fa095c77cfd3318cfdd1e7f6595d656d