test:qemu provide the right path for config
When krtd starts it uses default_config.json to
get the config. If not config is given in commamd
line it will look for that file in the current
dir. qemu_rt starts the krtd from /tmp/reg-1 and
so use the -c option to provide the path of config file

Change-Id: I252f80e73dba4cea2a798195be3016e0a8810f04