test:qemu_rt increase qemu vm memory
With some tests we see kboin_rt sometimes get
killed because of out of memory. The total
memory used by kbin_rt comes to arounf 1.27G
The mapped bar memory is close to 1G and
270M is used by kbin_rt. This seems to be huge
and needs to be debugged/accounted for. Till
that done, the vm memory is increased so that
the test cases pass.

Change-Id: Ie7577585ec4d577f7a5f642ad24f854ec5330aea