[kaena-827] nn_executor: Specify node names for input files
With this, you can now specify the name of the input node that each
input file goes to. For 2+ inputs, the names are required to avoid
dependency on the order, which is non-deterministic (e.g. --input_files
in_1=file1.npy in_2=file2.npy).

Multiple files without node names are still supported for backward
compatilbity to support batching, but this requires the nn_graph.json
not to define any inputs.

Change-Id: I76c706566a6fa4bfb42cca8a15a1e29b6c22e5b7