test:qemu_rt add pool option for emulator
Currently for the emaulator in each test case
lots of time is spend on starting testbench/vm.
With pool mode, this process is split into following
phases.
- start the pool
- start the RunAll
- clean the pool
This way when the RunALL is running it just keeps
restaring the krtd between test cases

grpc is used by krtc to talk to the emulator and that
is hard coded for the time being, since we can have
only 1 user per emulator to actively work.

Change-Id: I46eac369d5700ea61ff24040e34bd0cdcdffcad7