[SeaLife] Jenkins: Add correct handling of submodules
We're not currently able to handle changes to the submodule structure of
the TVM repo. A recent patch does just that but cannot pass because
submodules are not updated.

With this, we update the submodules after applying the gerrit patch and
clean any left over (old) submodules.

We also now try to apply gerrit patches to any submodule without naming
them explicitly so their names/locations can change without requiring an
update ot the Jenkinsfile.

Change-Id: Ib42b7005b848921c2386732306894b290b07cf88
Signed-off-by: Tobias Edler von Koch <kochtobi@amazon.com>