test:qemu_rt add multiple inference of same input
Add the knob to do multiple inferences. This will
only inference with the same input file.
The first inference outputs SHA are saved and
subsequent inference outputs are compared with the SHA.
If they differ return failure. If all inferences
produce the same output then the output is given
back to nn_executor which compares with the expected
output

Change-Id: Iad783501635fdb1f27802a8594a6bfa5f3bc26ff
Signed-off-by: vrnatham <vrnatham@amazon.com>