{WIP} test:jenkins create jenkinsfile per pipeline
Currently there are some stages that are duplicate
between all these files. Going forward will find
a way to modularize these. Each of the jenkisnfile
defined here should correspond to a pipeline and any
additions/deletions of stage/test to the pipeline
needs to be controlled using the jenkinsfile
These files are added for test purposes only
for the time being and will be tested in the
backup jenkins server

Change-Id: I749b5281e83485e965f54256eaa5c553f3f35e75
Signed-off-by: vrnatham <vrnatham@amazon.com>