test:qemu_rt remove skip test from xml result
Lots of non compiler tests are skipped for valid
reasons like test input is not supported/some tests
are specific only to emu etc. In the resuls these show
up as skipped and can be confusing. Since don't know
of an easy way to not count this tests using pytest
adding this hack here so that jenkins dashboard does
not show huge number of skipped tests

Change-Id: I270f2c2150ff3492077e9cd12542b7e7725eec8b
Signed-off-by: vrnatham <vrnatham@amazon.com>