test: jenkins dont bail out when no junit result
In jenkins rume sometimes test cases timeout and that
extends the total time and so might have to kill the
jenkins job. In those cases RunAll does not generate
the junit result (this needs to be fixed in runall too).
In jenkins atleast move to the next step and create
artifacts so that we can debug

Change-Id: I26b4a89823fea922154b203f8033f61f46931b14
Signed-off-by: vrnatham <vrnatham@amazon.com>