kaena-1130: reorder MatMuls in uint8 performance mode
If uint8 performance mode is requested, whenever we accumulate
a sequence of MatMul waveops in PSUM due to a non-1x1 filter,
always place the "center" MatMul waveop at the first place
and so it will mark the entire PSUM region that will be operated
by this sequence of MatMuls as "accumulating". This is to avoid
the weird accumulation behavior that happens when a MatMul puts
8 bytes at, say, 0x0, where 0x0 is not written by any ealier
MatMul and so anything being written to 0x0 is treated as a new
value, but 0x4 is in fact already written by some other MatMul
and so things being written to 0x4 should be treated as a value
that is going to be accumulated with the existing value at 0x4.

Change-Id: I5f517be84ce8f7ed99e0d4a779e8f72b912d40e6
Signed-off-by: Haichen "HC" Li <lhaiche@amazon.com>