Generate proper release version
We are adding a release process which generates git tags for each repo.
The tag is in the format of v{major}.{minor}.{build}

Each time a release is triggered, the build number will be incremented.

We will manually change the major/minor version by branching off old
versions and manually tagging the new master branch.

To this end, use `git describe --abbrev=0` to get the most recent tag
and use it for the package name.

Change-Id: I823dde1225ad63ed2c8336be35dc145c5740c63e
Signed-off-by: Joe Zhou <joezho@amazon.com>
Reviewed-on: https://siopt-294228977.us-east-1.elb.amazonaws.com/review/11223
Reviewed-by: Samuel Jacob <samjaco@amazon.com>
Tested-by: Jenkins